@TomBayne 2025-04-18
This could probably use some more work. In all my tests it demonstrated a clear performance increase, but it seems like this functionality is way slower than it needs to be.

I may look more into this in the future, but for now this is an improvement.